Add information about the game type and commit hash to the game (#6393)
## Description of the proposed changes

Embed the hash and game type into the game by processing the files just
before we deploy them. For now we can log this information to make it
easier to understand for the developerat what commit a replay was. In
the future we can use the same data to inform the host that a new game
version is available and that he should restart.

## Testing done on the proposed changes

I ran the bash scripts manually by copying them into a bash interpreter.

## Additional context

Users always experience a (short) duration of desyncs when a release
happens. And in general, users are unaware of the exact version that
resides on the FAF Develop and FAF Beta Balance game types