Prevent simultaneous deployments across all branches (#6400)
## Description of the proposed changes

By experience I know that multiple deployments in quick succession tend
to not end up deploying. A commit to any of the `deploy/` branches is
sufficient to trigger the deployment. The mistake is trivial to make as
I just did it by accident:


![image](https://github.com/user-attachments/assets/1fc076e2-fc1f-41ff-b517-4d51c895bb8e)

Which resulted in:


![image](https://github.com/user-attachments/assets/137279e5-4ee1-41c9-a6cc-1cc2889e4c94)

With these changes the workflows should cancel when they detect there
are concurrent workflows.