Add new general purpose issue template and reorder templates (#6421)
Our issue templates lack a general purpose form that covers for example
changes to existing features. That is not a bug report and not really a
feature request. So I added a general purpose template. I also reordered
the templates to have the specific and probably rarely used ones at the
bottom. (Assuming the templates show up in alphabetical order of the yml
files)