Unify the MuzzleVelocity stat of the Aeon Cruiser's missiles (#6440)
The Infinity Class Cruiser is equipped with two copies of Zealot AA
Missile launchers, both of which target the same unit. These two weapons
function identically except for the `MuzzleVelocity` stat, where one
weapon has a higher value than the other (though the left weapon also
does not have a `TargetCheckInterval`, and prefers the target of the
right weapon). This is misleading because both unit databases show the
AA batteries as having a `MuzzleVelocity` of `40`, when in reality, this
is only true for one of the launchers. The models on the unit also look
identical.

I am not aware of any other unit that does this; the Keefer Class has a
MuzzleVelocity of 30 for both of its AA batteries. The UI also does not
indicate a difference in MuzzleVelocity between both weapons.

**Infinity Class: T2 Cruiser (UAS0202):**
   - Zealot AA Missile (right launcher):
      - MuzzleVelocity: 40 --> 35
   - Zealot AA Missile (left launcher):
      - MuzzleVelocity: 30 --> 35