Various improvements for the Pulsar (#6423)
## Description of the proposed changes

**Pulsar: T3 Mobile EMP Missile Launcher (SRL0310):**
  - Categories added:
    - `PRODUCTDL`
    - `SNIPEMODE`
  - Othismash EMP Missile Barrage:
    - DamageRadius: 0 --> 1 (same as its EMP weapon)
- TurretPitchRange: 15 --> 40 (required against nearby units and units
on top of mountains)
- TurretPitchSpeed: 20 --> 50 (improves the responsiveness of the
turret)
- Introduce a taller firing arc to allow the Pulsar to shoot over
obstacles more easily

## Explanation of the changes

- The `PRODUCTDL` category is added as the Pulsar is content added
outside of the main game and the expansion.
- The `SNIPEMODE` category is added as it is instrumental for using the
Pulsar against ACU's equipped with personal shields. The Absolver also
has this category.
- Its weapon gains the ability to deal damage in an AoE, differentiating
it more from the Absolver. The effects of the missiles appear like they
deal AoE damage and the weapon's EMP effect also has a radius of 1.
- The `TurretPitchRange` is increased so it can hit units close to its
model. Right now, this is not the case; its missiles will overshoot
closer units. The model does provide enough leeway to make this change.
- `UseFiringSolutionInsteadOfAimBone = true`; required to fix
overshooting.
- The ballistic arc was very shallow due to `TrackTarget` being set to
`true` in the projectile's blueprint.

## Checklist
- [x] Changes are documented in the changelog for the next game version

---------

Co-authored-by: FAForever <administrator@faforever.com>