Improve the visual representation of multiple projectiles by increasing their `StrategicIconSize` (#6465)
## Description of the proposed changes
**Usha-Ah: Seraphim T3 Sniper Bot (secondary firing mode):**
The current `StrategicIconSize` of this projectile is set to `1`. This
is too small, because the projectile is very fast and thereby easy to
miss. A `StrategicIconSize` of `1` also does not represent the actual
strength of the projectile well (2000 damage).

I changed the `StrategicIconSize` to `4` in this PR. Alternative and
possibly preferable options could be:
- Only increase it to `3` since `4` is only used by TML missiles.
- Decrease the `StrategicIconSize` of the first firing mode to `2` and
and only increase the size of the second firing mode to `3`.

**Othuum: T3 Siege Tank (heavy cannon); Yenzyne: T2 Hover Tank:**
StrategicIconSize: 1 --> 2

**Percival: T3 Armored Assault Bot:**
StrategicIconSize: 2 --> 3

## Checklist
- [x] Changes are documented in the changelog for the next game version