Refactor of `maputils.lua` (#6495)
## Description of the proposed changes

Updates the annotations surrounding the map utilities module and
decouples some of the functionality. This allows the logic in other
modules, such as the (auto) lobby, to be simplified.

Removes the map blacklist module. It has not been updated for a decade
and these days we do have decent version control of maps in the vault.

## Testing done on the proposed changes

Tested in both the regular and the coop lobby.