Shader rework (#6485)
Now every shader is capable of using the new, better way of calculating
the water absorption. The only requirement is that the light multiplier
is bigger than 2.1. This is good, because the mesh shader doesn't know
what terrain shader is in use and has worked like this before. This
could lead to unfitting results, if the light multiplier was big enough,
but a terrain shader was used that only supported the legacy water
calculation. This is now fixed.

I also made the decals able to use pbr light calculations. Now they
behave consistently like the ground they are on. (Also the map editor
does this automatically and we can't prevent it, so it helps the editor
produce results that are consistent with the game.)
Splats are unaffected for now.

One thing about the decals is that they theoretically have a texture to
define the specularity (or roughness), but this texture was black for
all decals that I tested. I don't know if there are any decals that have
this defined.