Enable the SR's Iridium Rocket Pack to deal damage in an area of effect (#6383)
## Description of the proposed changes
The SR has four air-to-ground weapons: two bolters, which are the more
dominant weapons, and two rocket packs, which are often overlooked. In
the current release, only the bolters have area of effect damage, the
rocket packs have none at all.

From a design perspective, this is suboptimal since the rockets have a
decently large explosion effect. The Renegade uses the same projectile
with similar effects and has an area of effect radius of 3. It is also
not intuitive, because it could be reasonably assumed that the rocket
packs perform similarly to either the Renegade or the other
air-to-ground weapons of the SR. Balance wise, it results in the rocket
packs underperforming against smaller targets, such as Titans.

**Soul Ripper: Experimental Gunship (URA0401)**:
   - Iridium Rocket Pack (x2):
      - DamageRadius: 0 --> 3
      - Damage: 190 --> 150
      - DPS: 285 --> 225

## Additional context
Since this does increase the effectiveness of the rocket packs, their
DPS is reduced.

## Checklist
- [x] Changes are documented in the changelog for the next game version