Fix Seraphim carrier/T3 Sub and Cybran carrier AA missing their targets (#6427)
## Reasoning for the changes
The Seraphim carrier hits air targets less reliably than it should. It has the same weapon as the Seraphim SAM
(by name and visuals), but doesnt have any of the expected tracking
properties. The muzzle velocity and pitch range is also slightly
different.
Cybran carrier's projectile has tracking which turns off the weapons'
target leading. The tracking is not strong enough for the projectiles to
correct their aim with the high velocity of the projectiles. The Cybran
carrier actually has a worse hit rate than the Seraphim carrier because
of the combination of turning off leading and too weak tracking.

The Seraphim T3 subhunter does not get any tracking as the balance team decided the anti air is already quite strong and does not need further buffs as it is not the main purpose of the unit.

## Description of the proposed changes
Copy the Seraphim SAM projectile physics to the carrier
projectile. Also copy it to the Cybran carrier's
projectile to improve its hit rate.
Copy the Seraphim SAM muzzle velocity and turret pitch to the carrier so it can aim straight upwards.