Fix the launch event of a Billy being synced to the UI of enemies/neutrals(#6546)
The launch event would not show by default for enemy/neutral players. However, if the information is available to the UI then all it takes is a single rogue UI mod to override the logic and show it anyway. Now the data is simply not available in the UI for enemy and neutral players.