Fix Sera dummy weapon increasing turret yaw speed
The yaw speed was being set equal to the unit turn speed (90), while the actual desired stats are 70/50 yaw speed, so this is a very significant change.

In terms of commit history, the 90 yaw speed dummy weapon came in a game change (aug 2017) before the balance changes to the desired yaw speeds (jan 2018).