Buff UEF ACU Drone's rebuild costs (#6591)
UEF ACU drone rebuild costs
Mass: 160 -> 60
Energy: 1600 -> 800
Buildtime: 400 -> 200
- The drone build time is now affected by the ACU's buildpower.

- After testing the rebuilding during a stall, it seems like
`CreateEconomyEvent` works in blocks of buildtime in the same way as
silos. Reworking the drone rebuild process to be the same as the
kennel's (`TPodTowerUnit`) would fix this.
- The Drone auto-rebuild toggle is still using the very old gpg system
of removing the drone enhancement when the drone dies. To implement it
correctly there would have to be some UI data set on the ACU to create a
"rebuild drone" button when the drone is destroyed but not rebuilding,
and then that button would issue a sim callback or script task order
that initiates the drone rebuild sim-side.