Fix aeon sniper walk animation disrupting aim
The walk animation moves the sniper up and down a lot, which throws off the aim because the engine tells the sniper to fire when it is aimed correctly, but the muzzle charge delay delays that firing, at which moment the weapon may not be aimed anymore. Increasing the pitch speed compensates for that up/down animation so that the sniper is aimed every tick, making the delay not a problem.