Revert increased size of Seraphim air transport footprints (#6755)
This re-introduces an issue with them not moving off building locations, but fixes a more important issue of them not wanting to land on certain plateaus.
An ideal fix would move the model around so that it can have a footprint more centered on its hitbox/where units are attached.