Fix cancelling tempest firing cycle
MuzzleChargeDelay is bad because it desyncs the projectile creation and the projectile firing from the engine when the weapon is aimed.
We ignore the engine's behavior by instead basing the reload on rack salvo reload time, and setting the fire rate low. This allows us to cancel firing during the charging sequence without firing and without triggering the reload delay.