Fix replay ID extraction possible breaking UI
This does not affect FAF games an the previous code works, when the format of the `savereplay` is exactly as expected.

But for whoever would messing around with saving the replays only locally. It could end up trying to make arithemic operation on with nil. And that breaks the whole UI as this function is called from userInit.lua

This new version should work on all string safely, while grabbing the replay ID correctly from the specified format.