Fix wrong army report to lobby server with uneven teams
Army IDs are same as the slot indexes only when the teams are the same. Not much a problem with now with FAF matchmaker, as the teams are always the same size.

But if in the future (GW or different queue) wants to have uneven teams. This code will support it to send correct army IDs.

This is how it works in custom lobbies