Fix ship wake effects disappearing after a while
Fixes #6980
When ships start moving the 'Default' terrain type is used for picking the effects, which has Wake effects defined.

But as the ship is moving it encounters different WaterXX terrain types and they dont specify any extra affects, so the 'Default' one should be used.
That's what this change fixes.

As a side effect it might fix some other missing effects, since the 'Default' terrain type is suppose to be used when the non default one doesnt specify the effects.