Fix removal of invalid squad templates (#6984)
When the platoon template is process, some squad units can end up with quantity 0 and those should be removed.

table.remove shift all remaining values down, which the old code didnt consider and it was skipping the more squads the more got removed.

The fix iterates the table from the back which makes table.remove safe to use